Introduce validator struct
