+++ schemas.go
@@ -1,6 +1,8 @@
 package schemas
 
-import _ "embed"
+import (
+	_ "embed"
+)
 
 //go:embed step.schema.json
 var StepSchema string
